updated the folder structure and imports
