+++ src/components/Main.js
@@ -1,18 +1,18 @@
+import { useEffect } from "react";
+
 import Navbar from "@/components/navbar/Navbar";
 import Grid from "@/components/hero/Grid";
 import Footer from "@/components/footer/Footer";
 import NewGame from "@/components/gameSetup/NewGame";
 import GameEnd from "@/components/gameEnd/GameEnd";
-import { useEffect } from "react";
-import useGameStore from "./gameStore";
+
+import useGameStore from "@/store/gameStore";
 
 const Main = () => {
   const settings = useGameStore((store) => store.settings);
   const adjustGridSize = useGameStore((store) => store.adjustGridSize);
   const adjustPlayers = useGameStore((store) => store.adjustPlayers);
   const shuffleGridValues = useGameStore((store) => store.shuffleGridValues);
-  const totalScore = useGameStore((store) => store.totalScore);
-  const gridSize = useGameStore((store) => store.grid.gridSize);
 
   useEffect(() => {
     adjustGridSize();
@@ -30,11 +30,9 @@ const Main = () => {
     <>
       <NewGame />
       <Navbar />
-      <div className="lg:container mx-auto px-4">
-        <Grid />
-      </div>
+      <Grid />
       <Footer />
-      {totalScore * 2 === gridSize * gridSize && <GameEnd />}
+      <GameEnd />
     </>
   );
 };
