+++ src/components/gameEnd/Card.js
@@ -1,5 +1,5 @@
 import React from "react";
-import useGameStore from "../gameStore";
+import useGameStore from "../../store/gameStore";
 
 const Card = ({ details }) => {
   const { title, subtitle, data } = details;
@@ -8,6 +8,20 @@ const Card = ({ details }) => {
   );
   const restartGame = useGameStore((store) => store.restartGame);
   const shuffleGridValues = useGameStore((store) => store.shuffleGridValues);
+  const toggleGameEnd = useGameStore((store) => store.toggleGameEnd);
+
+  const handleRestart = () => {
+    restartGame();
+    shuffleGridValues();
+    toggleGameEnd();
+  };
+
+  const handleNewGame = () => {
+    restartGame();
+    shuffleGridValues();
+    changeSettingsStatus();
+    toggleGameEnd();
+  };  
 
   return (
     <div
@@ -54,20 +68,13 @@ const Card = ({ details }) => {
         <div className="flex flex-col md:flex-row gap-4 md:gap-[0.875rem] md:pt-4 text-4.125 md:text-5">
           <button
             className="basis-full text-lg md:text-xl font-bold px-4 lg:px-6 py-2 lg:py-3 rounded-full bg-[#fca516] hover:bg-[#fcba4f] text-white"
-            onClick={() => {
-              restartGame();
-              shuffleGridValues();
-            }}
+            onClick={() => handleRestart()}
           >
             Restart
           </button>
           <button
             className="basis-full text-lg md:text-xl font-bold px-4 lg:px-6 py-2 lg:py-3 rounded-full bg-[#dfe7ec] hover:bg-[#6393b6] text-[#32485a] hover:text-white"
-            onClick={() => {
-              restartGame();
-              shuffleGridValues();
-              changeSettingsStatus();
-            }}
+            onClick={() => handleNewGame()}
           >
             Setup New Game
           </button>
